Add merge data test
